putnbr

+++ Libft/ft_strjoin.c
@@ -10,10 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
 #include <stdio.h>
 #include <stdlib.h>
 
-int	ft_strlen(char const *str)
+size_t	ft_strlen(char const *str)
 {
 	int	i;
